asdf

+++ final/part1.cpp
@@ -18,7 +18,7 @@ int main (void) {
     input.open("tf.txt");
 
     while (getline(input, temp)) {
-        size_t numPos = temp.find("question");
+        unsigned numPos = temp.find("question");
         string num = temp.at(numPos);
         size_t pos = temp.find("is");
         string end = temp.substr(pos + 3);
